minimal/requirements/base.txt
Django==1.8.12
Pillow==2.9.0
argparse==1.2.1
django-authtools==1.4.0
psycopg2==2.6.1
django-extensions==1.6.1
ipdb==0.8.2
django-dbbackup==2.3.3
django-braces==1.8.1